+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -42,8 +42,8 @@ What kind of change does this Pull Request introduce?
 - Get the code
 
 ```
-git clone git@github.com:ks6088ts/template-python.git
-cd template-python
+git clone git@github.com:ks6088ts-labs/sandbox-gpt-researcher.git
+cd sandbox-gpt-researcher
 ```
 
 - Test the code
