+++ .github/workflows/ghcr-release.yaml
@@ -4,7 +4,7 @@ on:
     tags:
       - "v*"
 env:
-  IMAGE_NAME: template-python
+  IMAGE_NAME: sandbox-gpt-researcher
 jobs:
   ghcr:
     runs-on: ubuntu-latest
