+++ README.md
@@ -1,9 +1,9 @@
-[![test](https://github.com/ks6088ts/template-python/actions/workflows/test.yaml/badge.svg?branch=main)](https://github.com/ks6088ts/template-python/actions/workflows/test.yaml?query=branch%3Amain)
-[![docker](https://github.com/ks6088ts/template-python/actions/workflows/docker.yaml/badge.svg?branch=main)](https://github.com/ks6088ts/template-python/actions/workflows/docker.yaml?query=branch%3Amain)
-[![docker-release](https://github.com/ks6088ts/template-python/actions/workflows/docker-release.yaml/badge.svg)](https://github.com/ks6088ts/template-python/actions/workflows/docker-release.yaml)
-[![ghcr-release](https://github.com/ks6088ts/template-python/actions/workflows/ghcr-release.yaml/badge.svg)](https://github.com/ks6088ts/template-python/actions/workflows/ghcr-release.yaml)
+[![test](https://github.com/ks6088ts-labs/sandbox-gpt-researcher/actions/workflows/test.yaml/badge.svg?branch=main)](https://github.com/ks6088ts-labs/sandbox-gpt-researcher/actions/workflows/test.yaml?query=branch%3Amain)
+[![docker](https://github.com/ks6088ts-labs/sandbox-gpt-researcher/actions/workflows/docker.yaml/badge.svg?branch=main)](https://github.com/ks6088ts-labs/sandbox-gpt-researcher/actions/workflows/docker.yaml?query=branch%3Amain)
+[![docker-release](https://github.com/ks6088ts-labs/sandbox-gpt-researcher/actions/workflows/docker-release.yaml/badge.svg)](https://github.com/ks6088ts-labs/sandbox-gpt-researcher/actions/workflows/docker-release.yaml)
+[![ghcr-release](https://github.com/ks6088ts-labs/sandbox-gpt-researcher/actions/workflows/ghcr-release.yaml/badge.svg)](https://github.com/ks6088ts-labs/sandbox-gpt-researcher/actions/workflows/ghcr-release.yaml)
 
-# template-python
+# sandbox-gpt-researcher
 
 This is a template repository for Python
